ENH: Added meteo processes
Added a number of (hydro-)meteorological processes, especially needed for calculation of evapotranspiration. Also included a separate file of mathematical and physical constants. To test or illustrate the output of the routines some R files have been added as well as some pdf files that have been used as source of literature.