Might be relevant: Treatment of 'totalheat' in et_pot() and et_act()
Variable 'totalheat' (heat conduction into soil AND plants due to physical and biochemical energy storage) set equal to soil heat flux if the latter was given as input and the former not. At the moment this variable is har to interpret and quantify so that should be more or less adequate.