ENH: Added test analysis for soil water processes
Test analysis based on class test_soilwat. It includes a simple ECHSE set-up with a test soil parameterisation and analysis, including graphical outputs visualizing the bahaviour of the soil water processes included in ECHSE. It further includes comparisons with Hydrus-1D.

This test set-up can be used to understand ECHSE's soil water processes and to test and investigate future developments.